crewai-content-creator/requirements.txt
crewai==0.102.0
langchain==0.3.19
langchain_together==0.2.0
langchain_groq==0.2.2
langchain-exa==0.2.0
matplotlib==3.9.4
langchain_openai==0.2.14
langchain_experimental==0.3.4
python-dotenv
pypdf
requests
streamlit
langchain-community
streamlit-option-menu
pyperclip
pre-commit
tweepy
yfinance
pandas
numpy
playwright
reportlab
markdown
bs4
langfuse
openlit
crewai_tools